fixup: handling of multi-byte characters
The type of screen buffer is wchar_t for every cell.
Some multi-byte characters need multi columns to use, so
we should fill "No use". for example, L'\0'.

In the previous implementation, multibyte judgment was done every read,
so it can not be judged correctly when read comes back in the middle of
multibyte.

Signed-off-by: Yuichi Nakai <xoxyuxu@gmail.com>